requirements.txt
dash==1.9.0
dash-bootstrap-components==0.8.3
dash-core-components==1.8.0
dash-daq==0.3.1
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.0
EMD-signal==0.2.10
fastdtw==0.3.4
joblib==0.15.1
matplotlib==3.0.3
missingno==0.4.2
numpy==1.18.4
pandas==1.0.3
plotly==4.8.1
PyYAML==5.1.1
jupyter-dash==0.3.1
scikit-learn==0.23.1
scipy==1.5.0
seaborn==0.10.1
statsmodels==0.10.1
